Add support for Proxy Protocol.
This is useful for Services that are fronted by little
proxy, but do not want the client's ip address to be
obscured because the server's business logic involves
using the client's ip address.
https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt

The feature is controlled by two options in the
DefaultHttpProxyServer class :
1) sendProxy - to send a proxy protocol header to server
2) acceptProxy - to accept one from client
Based on these configurations, a proxy protocol message
decoder and encoder is added in ClientToProxyConnection
and ProxyToServerConnection classes respectively.
The default values of these configurations is false.

Tests :
Existing tests passed.
Unit tests added for send, accept and relay proxy header